+++ test/plugins/test_discogs.py
@@ -409,7 +409,9 @@ class DGAlbumInfoTest(BeetsTestCase):
         )
         d = DiscogsPlugin().get_album_info(release)
         assert d.artist == "ARTIST NAME & OTHER ARTIST"
+        assert d.artists == ["ARTIST NAME", "OTHER ARTIST"]
         assert d.tracks[0].artist == "TEST ARTIST"
+        assert d.tracks[0].artists == ["TEST ARTIST"]
         assert d.label == "LABEL NAME"
 
     def test_strip_disambiguation_false(self):
@@ -448,35 +450,62 @@ class DGAlbumInfoTest(BeetsTestCase):
         )
         d = DiscogsPlugin().get_album_info(release)
         assert d.artist == "ARTIST NAME (2) & OTHER ARTIST (5)"
+        assert d.artists == ["ARTIST NAME (2)", "OTHER ARTIST (5)"]
         assert d.tracks[0].artist == "TEST ARTIST (5)"
+        assert d.tracks[0].artists == ["TEST ARTIST (5)"]
         assert d.label == "LABEL NAME (5)"
         config["discogs"]["strip_disambiguation"] = True
 
 
 @pytest.mark.parametrize(
-    "track_artist_anv,track_artist",
-    [(False, "ARTIST Feat. PERFORMER"), (True, "VARIATION Feat. VARIATION")],
-)
-@pytest.mark.parametrize(
-    "album_artist_anv,album_artist",
-    [(False, "ARTIST & SOLOIST"), (True, "VARIATION & VARIATION")],
-)
-@pytest.mark.parametrize(
-    "artist_credit_anv,track_artist_credit,album_artist_credit",
+    "track_artist_anv,track_artist,track_artists",
     [
-        (False, "ARTIST Feat. PERFORMER", "ARTIST & SOLOIST"),
-        (True, "VARIATION Feat. VARIATION", "VARIATION & VARIATION"),
+        (False, "ARTIST Feat. PERFORMER", ["ARTIST", "PEFORMER"]),
+        (True, "VARIATION Feat. VARIATION", ["VARIATION", "VARIATION"]),
+    ],
+)
+@pytest.mark.parametrize(
+    "album_artist_anv,album_artist,album_artists",
+    [
+        (False, "ARTIST & SOLOIST", ["ARTIST", "SOLOIST"]),
+        (True, "VARIATION & VARIATION", ["VARIATION", "VARIATION"]),
+    ],
+)
+@pytest.mark.parametrize(
+    (
+        "artist_credit_anv,track_artist_credit,"
+        "track_artists_credit,album_artist_credit,album_artists_credit"
+    ),
+    [
+        (
+            False,
+            "ARTIST Feat. PERFORMER",
+            ["ARTIST", "PEFORMER"],
+            "ARTIST & SOLOIST",
+            ["ARTIST", "SOLOIST"],
+        ),
+        (
+            True,
+            "VARIATION Feat. VARIATION",
+            ["VARIATION", "VARIATION"],
+            "VARIATION & VARIATION",
+            ["VARIATION", "VARIATION"],
+        ),
     ],
 )
 @patch("beetsplug.discogs.DiscogsPlugin.setup", Mock())
 def test_anv(
     track_artist_anv,
     track_artist,
+    track_artists,
     album_artist_anv,
     album_artist,
+    album_artists,
     artist_credit_anv,
     track_artist_credit,
+    track_artists_credit,
     album_artist_credit,
+    album_artists_credit,
 ):
     """Test using artist name variations."""
     data = {
@@ -558,13 +587,21 @@ def test_anv_album_artist():
     config["discogs"]["anv"]["artist_credit"] = False
     r = DiscogsPlugin().get_album_info(release)
     assert r.artist == "ARTIST"
+    assert r.artists == ["ARTIST"]
+    assert r.albumartist == "ARTIST"
+    assert r.albumartist_credit == "ARTIST"
+    assert r.albumartists == ["ARTIST"]
+    assert r.albumartists_credit == ["ARTIST"]
     assert r.artist_credit == "ARTIST"
+    assert r.artists_credit == ["ARTIST"]
     assert r.tracks[0].artist == "VARIATION"
+    assert r.tracks[0].artists == ["VARIATION"]
     assert r.tracks[0].artist_credit == "ARTIST"
+    assert r.tracks[0].artists_credit == ["ARTIST"]
 
 
 @pytest.mark.parametrize(
-    "track, expected_artist",
+    "track, expected_artist, expected_artists",
     [
         (
             {
@@ -600,18 +637,25 @@ def test_anv_album_artist():
                 ],
             },
             "NEW ARTIST, VOCALIST Feat. SOLOIST, PERFORMER, MUSICIAN",
+            ["NEW ARTIST", "VOCALIST", "SOLOIST", "PERFORMER", "MUSICIAN"],
         ),
     ],
 )
 @patch("beetsplug.discogs.DiscogsPlugin.setup", Mock())
-def test_parse_featured_artists(track, expected_artist):
+def test_parse_featured_artists(track, expected_artist, expected_artists):
     """Tests the plugins ability to parse a featured artist.
-    Initial check with one featured artist, two featured artists,
-    and three. Ignores artists that are not listed as featured."""
-    t, _, _ = DiscogsPlugin().get_track_info(
-        track, 1, 1, ("ARTIST", "ARTIST CREDIT", 2)
-    )
+    Ignores artists that are not listed as featured."""
+    artistinfo = {
+        "artist": "ARTIST",
+        "artist_id": "1",
+        "artists": ["ARTIST"],
+        "artists_ids": ["1"],
+        "artist_credit": "ARTIST",
+        "artists_credit": ["ARTIST"],
+    }
+    t, _, _ = DiscogsPlugin().get_track_info(track, 1, 1, artistinfo)
     assert t.artist == expected_artist
+    assert t.artists == expected_artists
 
 
 @pytest.mark.parametrize(
@@ -635,6 +679,29 @@ def test_get_media_and_albumtype(formats, expected_media, expected_albumtype):
     result = DiscogsPlugin.get_media_and_albumtype(formats)
 
     assert result == (expected_media, expected_albumtype)
+
+
+@pytest.mark.parametrize(
+    "given_artists,expected_info,config_va_name",
+    [
+        (
+            [{"name": "Various", "id": "1"}],
+            {
+                "artist": "VARIOUS ARTISTS",
+                "artist_id": "1",
+                "artists": ["VARIOUS ARTISTS"],
+                "artists_ids": ["1"],
+                "artist_credit": "VARIOUS ARTISTS",
+                "artists_credit": ["VARIOUS ARTISTS"],
+            },
+            "VARIOUS ARTISTS",
+        )
+    ],
+)
+@patch("beetsplug.discogs.DiscogsPlugin.setup", Mock())
+def test_va_buildartistinfo(given_artists, expected_info, config_va_name):
+    config["va_name"] = config_va_name
+    assert DiscogsPlugin().build_artistinfo(given_artists) == expected_info
 
 
 @pytest.mark.parametrize(